Update scripts for more years, cross-state processing
Updated syntax to more easily process 2000 and 2010 census geographies in addition to 2020 (all scripts)

Added code to download ZCTA-level population data (script 01)

Added code to join populated blocks that do not overlap with ZCTAs to nearest ZCTA if within 1000 meters (block is left unlinked if >1000m from a ZCTA) - script 02a.

Added code to update weights to account for ZCTA population split across states. Where population is split across states, the weights and population are updated to be that for the total ZCTA, instead of the previous state-level population. Code for comparing the ZCTA block aggregate and census population is also included (script 03)